--- a/python/pyspark/rdd.py
+++ b/python/pyspark/rdd.py
@@ -76,6 +76,7 @@ class PythonEvalType(object):
     SQL_SCALAR_PANDAS_ITER_UDF = 204
     SQL_MAP_PANDAS_ITER_UDF = 205
     SQL_COGROUPED_MAP_PANDAS_UDF = 206
+    SQL_SCALAR_ARROW_UDF = 207
 
 
 def portable_hash(x):