--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/PythonUDF.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/PythonUDF.scala
@@ -28,7 +28,8 @@ object PythonUDF {
   private[this] val SCALAR_TYPES = Set(
     PythonEvalType.SQL_BATCHED_UDF,
     PythonEvalType.SQL_SCALAR_PANDAS_UDF,
-    PythonEvalType.SQL_SCALAR_PANDAS_ITER_UDF
+    PythonEvalType.SQL_SCALAR_PANDAS_ITER_UDF,
+    PythonEvalType.SQL_SCALAR_ARROW_UDF
   )
 
   def isScalarPythonUDF(e: Expression): Boolean = {