### What changes were proposed in this pull request?
This proposes to extend data exchange options for vectorized UDF functions, by introducing the possibility to have plain Arrow serialization without conversions to Pandas. 
This PR proposes to introduce SCALAR_ARROW pandas_udf type as a step in that direction. As detailed in the test case below, the performance results are quite promising, in particular when processing data sets with numerical arrays, showing a speedup of 3x in the test case.

### Why are the changes needed?
Having the possibility to deploy vectorized Python UDFs that bypass the conversion to_pandas, can provide performance improvements (see test below) and extended flexibility (as the Arrow type system is broader than Pandas's) for operating data manipulation, in particular this is relevant for data sets including numerical arrays, possibly of variable length. While conversion to Pandas covers many cases already for vectorized Python UDF in a satisfying manner, the proposed improvement applies to certain special cases, notably of interested for scientific computing (such as in high energy physics), where performant processing of arrays of variable length is of paramount importance.

A test case shows a measured increase of performance of about 3x using the proposed approach, compared to the equivalent processing with pandas_udf (details below in the test section).  Processing of Arrow data in the test case was done via the "awkward arrays" library (https://github.com/scikit-hep/awkward-array).
In general, we can imagine that more specialized libraries using Arrow for data exchange can appear in the future to address computational needs in specific domains, possibly covering also cases of ML and DL libraries. It can be advantageous for Apache Spark to add the possibility to hook to such libraries, as proposed in the PR or with similar methods, for gains in performance and extra flexibility.

### Does this PR introduce any user-facing change?
This PR introduces an extension of pandas_udf types with a new SCALAR_ARROW type.

### How was this patch tested?
1. Manually tested:  a proposed "SCALAR_ARROW" pandas_udf test, detailed below, runs in 21 seconds vs. 62 seconds of the corresponding version with standard "pandas_udf SCALAR" code. This is close to a 3x speedup. The test is implemented using the awkward array library https://pypi.org/project/awkward/ for data processing, in alternative to Pandas:
```
bin/pyspark --master local[2]

df = sql("select cast(1.0 as double) col1, rand(42) col2, Array(rand(),rand(),rand()) col3 from range(1e8)")
df.printSchema()

root
 |-- col1: double (nullable = false)
 |-- col2: double (nullable = false)
 |-- col3: array (nullable = false)
 |    |-- element: double (containsNull = false)


from pyspark.sql.functions import pandas_udf, PandasUDFType
from pyspark.sql.types import *
import awkward

@pandas_udf(ArrayType(DoubleType()), PandasUDFType.SCALAR_ARROW)
def test_arrow(col1):
    b = awkward.fromarrow(col1.chunk(0))
    c = awkward.toarrow(b * b)
    return c 

// Note: following SPARK-28264 this can also be written as:
import pyarrow as pa
@pandas_udf(ArrayType(DoubleType()))
def test_arrow(col1: pa.ChunkedArray) -> pa.Array:
    b = awkward.fromarrow(col1.chunk(0))
    c = awkward.toarrow(b * b)
    return c 

import time
start = time.time()
df.withColumn('test', test_arrow(df.col3)).write.format("noop").mode("overwrite").save()
end = time.time()
print(end - start)
```

2. This is the reference code with pandas_udf SCALAR (this runs in 62 seconds vs. 21 seconds of the code above with SCALAR_ARROW implementation):
```
@pandas_udf(ArrayType(DoubleType()), PandasUDFType.SCALAR)
def test_pandas(col1):
  return col1*col1

import time
start = time.time()
df.withColumn('test', test_pandas(df.col3)).write.format("noop").mode("overwrite").save()
end = time.time()
print(end - start)
```

3. This is how the test workload can be implemented using Spark SQL (with higher order SQL functions). This test rund in 11 seconds
```
import time
start = time.time()
df.selectExpr("col1", "col2", "col3", "transform(col3, x -> x * x) as test").write.format("noop").mode("overwrite").save()
end = time.time()
print(end - start)
```

In conclusion the test workload runs 6x slower than Spark SQL with pandas_udf of type SCALAR, and 2x slower than Spark SQL with the proposed arrow-only serialization with SCALAR_PYARROW pandas_udf.

----
In addition, we collected and compared Flamegraph profiling of the Python code execution on the Python daemons, in the 2 Python UDF cases tested (SCALAR UDF and nthe proposed SCALAR_ARROW type). Results show that in the SCALAR case **an important amount of time is spent in the to_pandas conversion** operations and in array processing, the SCALAR_ARROW case optimizes processing of array data and avoid conversion to Pandas for array data, thus saving considerable time. 

This is a Flamegraph of workload on the Python daemon when running the pandas_udf SCALAR_ARROW:
![](https://issues.apache.org/jira/secure/attachment/12987761/Flamegraph_test_pandas_udf_SCALAR_ARROW.png)

This is a Flamegraph of workload on the Python daemon when running the pandas_udf SCALAR:
![](https://issues.apache.org/jira/secure/attachment/12987760/Flamegraph_test_pandas_udf_SCALAR.png)

Co-authors of this work:
Luca Canali @LucaCanali 
Lindsey Gray @lgray
Jim Pivarski @jpivarski