@HyukjinKwon 2019-12-07
> by introducing options to have plain Arrow serialization without conversions to Pandas

I think this is what Arrow should fix if it costs extra since that's what the project targets, rather then trying to work around within Spark side.