@jpivarski 2019-12-07
(Full disclosure: I'm a co-author.) Considering that Arrow is lower-level than Pandas, I would have thought that the Pandas API would have been built on top of an Arrow API, as a (very popular) special case.

The argument is being made in terms of performance because in our application we have to work around DataFrame-construction that we don't want and don't need, and it's a considerable runtime cost. But the argument could also be made in terms of a layered architecture. Wouldn't it make more sense for the pandas_udf to be an application built on an arrow_udf?

I just want to make sure that the architectural point isn't lost in discussions about performance.