@HyukjinKwon 2019-12-07
I think it's too late to discuss about switching it from arrow pandas. The counter argument is that arrow is too low level API to be exposed to end users.

Arrow targets exactly the same problem you mentioned - no overhead between communication among other systems. Why dont we deduplicate this efforts in Arrow rather then working around in Spark only?