@jpivarski 2019-12-07
I forgot to mention that Arrow provides a broader type system than Pandas—there are data that can be expressed in Arrow but not in Pandas. A simple example that we encounter all the time is (in [DataShape notation](https://datashape.readthedocs.io/en/latest/)),

```
N * {"one": var * int64, "two": var * int64}
```

That is, a dataset of `N` rows with two columns, `"one"` and `"two"`, that both contain variable-length lists of integers. If the variable-length lists don't have the same length in every row, you can't express it with a Pandas `MultiIndex`, and can only resort to Python lists in the cells (which defeats the purpose of a vectorized interface, not just for speed, but for functionality). The above could be expressed with two DataFrames with different `MultiIndexes`, but then two DataFrames can't be sent to the same Python process for comparison. In our work, the columns `"one"` and `"two"` represent dozens of particle types that need to be analyzed together (e.g. to find evidence of invisible particles that decayed into multiple observed particle types, each of which in a different variable-length list).

Pandas is one application that can use a subset of Arrow. Our application fits neatly into Arrow, but not Pandas.

I don't want to derail this—@LucaCanali made a specific pull request and that's the one under review. I spoke up because it sounded like it was being judged as a hack for performance, when really it's for functionality.