@HyukjinKwon 2019-12-08
Pandas DataFrame can be easily changed to Arrow table. If Arrow manages to get rid of the overhead you guys observed, problem solved, no?

Looks like we're adding a workaround for the issue in Arrow side. Also, currently all Spark types are being mapped (except MapType but theres already a PR to fix). So, there's no missing type mapping that we can take advantages of by switching to Arrow directly either.