@jpivarski 2019-12-08
> Pandas DataFrame can be easily changed to Arrow table.

While a Pandas DataFrame can easily be changed into an Arrow Table, many Arrow Tables can't be changed into Pandas DataFrames (without putting Python objects into the cells). The Arrow type system is broader than Pandas's: lossless conversions are possible in one direction only.

The types we need to use are in the set that Pandas can't handle, and that's why we don't use Pandas. However, the vectorized UDF from Spark though Arrow is great, and we want to use that.